✨ Changed the naming and approach to map/char disp
Before I was tracking what was next as a way to generate a map. I flipped that around and now am just tracking what is next and what has passed as the way to place the character in an empty board. This presents a couple of new problems, but much easier that building some crazy way to populate a map given several next tile numbers. Still need to figure out how to move backwards and stop rendering the paths that didn't occur. Basically, need to figure out how to use one character component and move that component with the logic I already have.