✨ Big area works
Character can move in 2d space according to rooms that are available. Arrows indicate which direction the char can go at each room. Next steps are to bring in path memory, possibly using the array of past moves. Style some sort of fog that gets pushed out of the way when the user moves in a direction.